requirements.txt
torch>=1.9.1
torchvision>=0.3
numpy
opencv-python
matplotlib
kornia>=0.6.11
GitPython>=3.1.0
